Solution to exercise 36
This is a game about the obstacles faced by a girl dog who was lost by her parents at the village market. You have to accompany her to surmount these obstacles and help her to get back home. 

Please remember to entre a girl dog's name as an argument along with the script name while running it in Python.